pyproject.toml
[project]
name = "sun-marke"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ipykernel>=7.1.0",
    "langchain>=1.2.6",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.1.0",
    "playwright>=1.57.0",
    "python-dotenv>=1.2.1",
    "selenium>=4.40.0",
    "sentence-transformers>=5.2.0",
    "setuptools>=80.10.1",
    "streamlit>=1.53.1",
    "undetected-chromedriver>=3.5.5",
    "unstructured>=0.18.27",
]
